fix #22: compile with the default Less `math` option
Less provides a `math` option
(http://lesscss.org/usage/#less-options-math) that it uses to determine
when it should treat slashes as division vs literal slashes. The option
defaults to `always` but we’ve been compiling the project with `math`
set to `parens`, which doesn’t process slashes as division unless
they’re contained within parentheses.

This decision to use the stricter `math` option has turned out to be
problematic because it wasn’t documented anywhere, so users have been
assuming the project should compile using Less out of the box. As a
result, several users were seeing compiler errors and thought there must
be an issue with the project itself.

In order to provide a more seamless, plug-and-play experience, we’ll
now develop and compile the project using the default `--math=always`
option, and simply escape literal `/` characters whenever they’re
needed. This way no configuration will be need by the user.

This issue was essentially fixed by #21, but in this commit we’re
converting the expanded properties added in that PR back to their
shorthand versions and escaping the slashes instead. This way the syntax
stays as close to 1:1 with the Sass version as possible.